fixed assumes
